% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norway_metadata}
\alias{norway_metadata}
\title{Norway Benthic Macrofauna Metadata}
\format{
A data frame with 101 rows and 1 column:
\describe{
  \item{Area}{A grouping factor with 5 levels indicating the specific geographical region.}
}
}
\source{
\doi{10.1111/anzs.12176}
}
\usage{
norway_metadata
}
\description{
Metadata containing geographical location assignments for the 101 sampled sites
in the Norway benthic community dataset.
}
\keyword{datasets}
