% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norway_pa}
\alias{norway_pa}
\title{Norway Benthic Macrofauna Presence-Absence Data}
\format{
A matrix with 101 rows (sites) and 809 columns (species).
}
\source{
\doi{10.1111/anzs.12176}
}
\usage{
norway_pa
}
\description{
A presence-absence matrix for 809 marine macrofauna species across 101 sites,
utilized as a baseline biological validation dataset.
}
\keyword{datasets}
