% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkd_metabolites}
\alias{pkd_metabolites}
\title{Polycystic Kidney Disease (PKD) Metabolites}
\format{
A matrix with 95 rows (samples) and 1554 columns (metabolites).
}
\source{
Derived from the urinary metabolomics cohort in Houske et al. (2023).
}
\usage{
pkd_metabolites
}
\description{
An auto-scaled, log-transformed intensity matrix of urinary profiles (metabolites) from
early-stage Polycystic Kidney Disease patients and healthy controls.
}
\keyword{datasets}
