% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkd_metadata}
\alias{pkd_metadata}
\title{Polycystic Kidney Disease (PKD) Metadata}
\format{
A data frame with 95 rows (samples) and 1 column:
\describe{
  \item{Group}{A factor indicating disease classification status (Ctrl vs PKD).}
}
}
\source{
Derived from the urinary metabolomics cohort in Houske et al. (2023).
}
\usage{
pkd_metadata
}
\description{
A dataset containing group categories for the Polycystic Kidney Disease (PKD)
cohort, used to demonstrate multivariate testing under biological heterogeneity.
}
\keyword{datasets}
