% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norway_metadata}
\alias{norway_metadata}
\title{Norway Benthic Macrofauna Metadata}
\format{
A data frame with 101 rows and 1 column (Area).
}
\usage{
norway_metadata
}
\description{
Metadata containing the 'Area' factor (1-5) for the 101 sites
in the Norway dataset.
}
\keyword{datasets}
